merge adjacent tokens with the same class, allow null class
whitespace is now no longer 'pulled into' the token next to it,
but put in class-less spans. parsers can return null to indicate
they don't want to assign a class to the consumed text. the js
parser now also does this for operators and punctuation, so
that the amount of joinable spans is much greater