Support editor themes
This normalizes the classes used by most of the modes
(reStructuredText and diff didn't fit in), and makes them all
reference a central style sheet (under theme/). It also adds a theme
option that can be set to change the theme, and defines a few
different themes (some based on old mode colors).