More reliable line-breaking detection
This hack is both heroic and completely terrible: it adds
regexps for the various browsers that are used to detect situations
where line-wrapping is influenced by span boundaries. It then uses
HTML as similar as possible to the actual display HTML when measuring
character offsets, and inserts <wbr> elements to compensate for
the differences that the introduction of the measuring span might
introduce.

It seems to work, better than the old situation at least, but the
assumption that browsers use the same line-breaking algorithm across
versions and platforms is bound to bite us.

(Feature-detecting the equivalent of these regexps is possible, but
expensive -- takes about 200 milliseconds on Chrome, more on other
browsers.)