[gfm mode] Fix: autolinks syntax. New: double bracket links.
Autolinks were not being displayed as such in the editor, changed
'linkhref' to 'link' in line 72 to fix it.

Added support for GFM double bracket links. '[[link]]' is now displayed
correctly as a link by the editor.

Added coloring support for latex formulas, usage of
- "$formula$",
- "$$formula$$",
- "\( formula \)", and
- "\[ formula \]"
will now be highlighted with the same coloring used for strings.