[merge addon] Add highlightDifferences option and method
Provides new option (highlightDifferences) and method highlightDifferences(bool)
to the MergeView plugin.

If true (or undefined) the original behaviour is maintained, which is to show
the differences (line highlighting and gutter graphics), if false then
the changes will not be shown.

Calling the method causes the visual change to occur 'immediately'

Signed-off-by: ciaranj <ciaranj@gmail.com>