[show-hint addon] Docs for #2228
add docs that when completions are object they can have a from: and to:
that will be used by pick.

Make sidebar in manual scrollable for small screen especially usefull
when `Programming API` is unfolded

seem to look nice on Chrome and Firefox Osx, (scrollbar do not appear,
except when scrolling on Chrome) but should maybe be tested on linux.