[soy mode] Add
See https://developers.google.com/closure/templates/ for documentation.
Closure Templates support text, HTML, JS and CSS code.

I'm not extremely happy with the indentation which is currently based
just on the Soy tags, not on the underlying code. Although this style
is possible and sometimes used, it's not the most common style.