[haxe mode] Treat types as expressions when parsing an expression
Closes #3513

Not 100% sure that this is the best way to deal with this - but if
you're in an expression, and encounter what you think is a type, it's
probably better to change your assumption and treat it as you would a
variable/constant (Because that's what it probably is) rather than to
forget everything.

I ran the test suite, and it didn't break anything, and the linter,
which was quiet.